+++ api/v1/app.go
@@ -137,6 +137,7 @@ func (a *App) RegisterRoutes(router *gin.Engine) {
 		workoutGroup.GET("/:id/map-preview", a.getWorkoutMapPreview)
 		workoutGroup.GET("/:id/media/:filename/preview", a.getWorkoutMediaPreview)
 		workoutGroup.GET("/:id/media/:filename", a.getWorkoutMediaOriginal)
+		workoutGroup.GET("/:id", a.getWorkout)
 		workoutGroup.PUT("/:id", a.updateWorkout)
 		workoutGroup.DELETE("/:id", a.deleteWorkout)
 		workoutGroup.GET("", a.listWorkouts)